+++ src/api/server.d.ts
@@ -1,0 +1,6 @@
+/**
+ * LiquidGov REST API Server
+ */
+declare const app: import("express-serve-static-core").Express;
+export default app;
+//# sourceMappingURL=server.d.ts.map